requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai==0.5.4
python-dotenv==1.0.1
boto3==1.34.102
six==1.16.0
cors==1.0.1